+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to Transaction.query
+- ids parameter to Transfer.query
 
 ## [0.6.0] - 2020-08-20
 ### Added